+++ NEWS.md
@@ -1,4 +1,4 @@
-# openxlsx (development version)
+# openxlsx 4.2.7.1
 
 * It's now possible to insert a hyperlinked image by passing a URL, relative or absolute file path, or mailto string to the new `address` parameter of `insertImage()`.